combined declarations
