--- internal/protcl/prot.go
@@ -1,61 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c)  2018 Kasun Vithanage
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package protcl
-
-import "fmt"
-
-const (
-	REP_SIMPLE_STRING = "+"
-	REP_INTEGER       = ":"
-	REP_BULKSTRING    = "$"
-	REP_ERROR         = "-"
-)
-
-const (
-	WRONGTYP = "WRONGTYP"
-	ERR      = "ERR"
-)
-
-type Reply interface {
-	Reply() string
-}
-
-type Err interface {
-	Err() ErrorReply
-}
-
-type Message struct {
-	Rep Reply
-	Err Err
-}
-
-type ErrorReply struct {
-	Prefix string
-	Err    string
-}
-
-func (rep *ErrorReply) Error() string {
-	return fmt.Sprintf("-%s %s\r\n", rep.Prefix, rep.Err)
-}
